requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.2
motor==3.4.0
python-dotenv==1.0.1
anthropic==0.28.0
pydantic==2.7.1
httpx==0.27.0